Nullable schema constraints (#5)
* Unspecified nullable hashes don't enforce subfield constraints

The primary purpose of this feature is to allow the schema to enforce
constraints on subfields of nullable hashes without making the parent hash
required, e.g.

  schema type: Strict::Hash do
    field :data, type: Strict::Hash | Strict::Nil, nullable: true do
      field :id, type: Strict::String
      field :type, type: Strict::String.enum('users')
    end
  end

The above is now a valid schema that will not raise invalid property errors if
data is null (unspecified). However, if the data hash is specified at all,
even with no subfields or only fields that don't match the schema, the
constraints on id and type will be checked with the input.

This allows us to e.g. sanitize relationships in JSON:API schemas, which are
null to indicate a removal, otherwise must match the schema.

Two new tests have been added to check both scenarios on a nullable hash
(specified and not specified).

* Add a test with data: null

* Test cleanup